normalize youtube summarization
