+++ apps/dashboard/pages/api/chains/query.ts
@@ -1,23 +1,17 @@
-import { ConstructionOutlined } from '@mui/icons-material';
 import Cors from 'cors';
 import { NextApiResponse } from 'next';
-import { z, ZodSchema } from 'zod';
+import { z } from 'zod';
 
-import ChatModel from '@chaindesk/lib/chat-model';
-import { ModelConfig } from '@chaindesk/lib/config';
 import {
   createLazyAuthHandler,
   respond,
 } from '@chaindesk/lib/createa-api-handler';
-import ytTool, { Schema } from '@chaindesk/lib/openai-tools/youtube-summary';
 import runMiddleware from '@chaindesk/lib/run-middleware';
-import splitTextByToken from '@chaindesk/lib/split-text-by-token';
-import generateSummary from '@chaindesk/lib/summarize';
 import { AppNextApiRequest } from '@chaindesk/lib/types';
 import validate from '@chaindesk/lib/validate';
 import YoutubeApi from '@chaindesk/lib/youtube-api';
-import zodParseJSON from '@chaindesk/lib/zod-parse-json';
-import { AgentModelName, LLMTaskOutputType } from '@chaindesk/prisma';
+import ytSummarize from '@chaindesk/lib/youtubeSummarizer';
+import { LLMTaskOutputType } from '@chaindesk/prisma';
 import { prisma } from '@chaindesk/prisma/client';
 const handler = createLazyAuthHandler();
 
@@ -40,7 +34,6 @@ export const queryFreeTools = async (
     case LLMTaskOutputType.youtube_summary:
       const Youtube = new YoutubeApi();
       const videoId = YoutubeApi.extractVideoId(url);
-      //TODO: get video name,  description date published
       const videoSnippet = await Youtube.getVideoSnippetById(videoId!);
 
       if (!videoId) {
@@ -61,43 +54,7 @@ export const queryFreeTools = async (
           ...(found.output as any)?.['en'],
         };
       } else {
-        const transcripts = await YoutubeApi.transcribeVideo(url);
-
-        const text = transcripts.reduce((acc, { text }) => acc + text, '');
-
-        const [chunkedText] = await splitTextByToken({
-          text,
-          chunkSize: ModelConfig[AgentModelName.gpt_3_5_turbo].maxTokens * 0.7,
-        });
-
-        const model = new ChatModel();
-
-        const result = await model.call({
-          model: ModelConfig['gpt_3_5_turbo'].name,
-          tools: [ytTool],
-          tool_choice: {
-            type: 'function',
-            function: {
-              name: 'youtube_summary',
-            },
-          },
-          messages: [
-            {
-              role: 'system',
-              content:
-                'You are a helpful assistant. Your task is generate a very detailed summary of a given text in a comprehensive, educational way.',
-            },
-            {
-              role: 'user',
-              content: `Youtube video transcript: ### ${chunkedText} ###`,
-            },
-          ],
-        });
-
-        const data = zodParseJSON(Schema)(
-          result?.completion?.choices?.[0]?.message?.tool_calls?.[0]?.function
-            ?.arguments as string
-        );
+        const data = await ytSummarize(url);
 
         await prisma.lLMTaskOutput.create({
           data: {
