+++ packages/lib/loaders/youtube-video.ts
@@ -3,7 +3,11 @@ import axios from 'axios';
 import { AppDocument } from '@chaindesk/lib/types/document';
 import { DatasourceSchema } from '@chaindesk/lib/types/models';
 
+import { ModelConfig } from '../config';
+import countTokens from '../count-tokens';
+import timeCodeToSeconds from '../timeCodeToSeconds';
 import YoutubeApi from '../youtube-api';
+import ytSummarize from '../youtubeSummarizer';
 
 import { DatasourceLoaderBase } from './base';
 
@@ -11,6 +15,7 @@ type DatasourceYoutubeVideo = Extract<
   DatasourceSchema,
   { type: 'youtube_video' }
 >;
+const maxTokens = ModelConfig.gpt_3_5_turbo.maxTokens * 0.7;
 
 export class YoutubeVideoLoader extends DatasourceLoaderBase<DatasourceYoutubeVideo> {
   async getSize(text: string) {
@@ -24,16 +29,20 @@ export class YoutubeVideoLoader extends DatasourceLoaderBase<DatasourceYoutubeVi
       throw new Error('Fatal: missing youtube url.');
     }
 
-    let docs = [];
+    let docs = [] as AppDocument[];
     try {
-      const transcripts = await YoutubeApi.transcribeVideo(url);
-      docs = transcripts.map(({ text, offset }) => {
-        return new AppDocument<any>({
-          pageContent: text,
-          metadata: {
-            source_url: `${url}&t=${Math.ceil(offset / 1000)}`,
-          },
-        });
+      const { chapters, thematics } = await ytSummarize(url);
+
+      chapters.map(({ summary, startTimecode }) => {
+        docs.push(
+          new AppDocument<any>({
+            pageContent: summary,
+            metadata: {
+              source_url: `${url}&t=${timeCodeToSeconds(startTimecode)}`,
+              tags: thematics,
+            },
+          })
+        );
       });
     } catch (err) {
       docs = [
@@ -56,7 +65,6 @@ export class YoutubeVideoLoader extends DatasourceLoaderBase<DatasourceYoutubeVi
           datasource_name: this.datasource.name,
           datasource_type: this.datasource.type,
           custom_id: this.datasource?.config?.custom_id,
-          tags: this.datasource?.config?.tags || [],
         },
       };
     });
